# feat!:  updates the HashNode and HashLeaf methods to return error instead of panic and refactors the code