--- a/hasher.go
+++ b/hasher.go
@@ -80,7 +80,8 @@ func (n *Hasher) Write(data []byte) (int, error) {
 		// check the format of the data
 		leftChild := data[:n.Size()]
 		rightChild := data[n.Size():]
-		if err := n.ValidateNodes(leftChild, rightChild); err != nil {
+		err := n.ValidateNodes(leftChild, rightChild)
+		if err != nil {
 			return 0, err
 		}
 		n.tp = NodePrefix