--- a/hasher_test.go
+++ b/hasher_test.go
@@ -235,7 +235,6 @@ func TestHashNode_ChildrenNamespaceRange(t *testing.T) {
 			if tt.wantErr {
 				assert.True(t, errors.Is(err, tt.errType))
 			}
-
 		})
 	}
 }