@staheri14 2023-03-23
benchstat results:
```
goos: darwin
goarch: arm64
pkg: github.com/celestiaorg/nmt
                          │ master.txt  │        nmt-err-handling.txt        │
                          │   sec/op    │   sec/op     vs base               │
ComputeRoot/64-leaves-10    38.30µ ± 2%   38.93µ ± 4%       ~ (p=0.089 n=10)
ComputeRoot/128-leaves-10   74.73µ ± 1%   75.29µ ± 1%  +0.75% (p=0.023 n=10)
ComputeRoot/256-leaves-10   152.1µ ± 3%   152.6µ ± 1%       ~ (p=0.579 n=10)
geomean                     75.79µ        76.48µ       +0.91%

```